implement feedback from Piotr
