--- extensions/walletProviders/evm-wallet/src/index.ts
@@ -1,1 +1,0 @@
-export * from './evm-wallet.extension'
