+++ extensions/walletProviders/evm/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@tatumio/evm-wallet",
+  "name": "@tatumio/evm-wallet-provider",
   "version": "1.0.1",
   "description": "EVM provider with local wallet operations",
   "author": "Tatum",
-  "repository": "https://github.com/tatumio/tatum-js/extensions/walletProviders/evm-wallet",
+  "repository": "https://github.com/tatumio/tatum-js/extensions/walletProviders/evm",
   "files": [
     "dist/"
   ],
