+++ extensions/walletProviders/evm/src/types.ts
@@ -1,0 +1,15 @@
+export type XpubWithMnemonic = {
+  xpub: string
+  mnemonic: string
+  derivationPath: string
+}
+
+export type EvmWallet = {
+  address: string
+  privateKey: string
+  mnemonic: string
+}
+
+export type EvmTxPayload = {
+  [key: string]: unknown
+}
